+++ packages/plugin-session-replay-browser/src/session-replay.ts
@@ -114,6 +114,8 @@ export class SessionReplayPlugin implements EnrichmentPlugin<BrowserClient, Brow
         maxPersistedEventsSizeBytes: this.options.maxPersistedEventsSizeBytes,
         maxSingleEventSizeBytes: this.options.maxSingleEventSizeBytes,
         sendTimeoutMs: this.options.sendTimeoutMs,
+        handleSendEvents: this.options.handleSendEvents,
+        handleFetchConfig: this.options.handleFetchConfig,
       };
 
       await this.sessionReplay.init(config.apiKey, this.srInitOptions).promise;