+++ packages/plugin-page-url-enrichment-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.15](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-url-enrichment-browser@0.7.14...@amplitude/plugin-page-url-enrichment-browser@0.7.15) (2026-06-30)
+
+**Note:** Version bump only for package @amplitude/plugin-page-url-enrichment-browser
+
+
+
+
+
 ## [0.7.14](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-url-enrichment-browser@0.7.13...@amplitude/plugin-page-url-enrichment-browser@0.7.14) (2026-06-25)
 
 **Note:** Version bump only for package @amplitude/plugin-page-url-enrichment-browser